GameManager, LapManager, PlayerLapCollision
GameManager.cs has 4 events -> "OnGameStart", "OnLapComplete", "AllLapsComplete", "OnGameLoose". It's also a singleton so it can be accesed from anywhere
PlayerCollisionTrigger.cs calls "OnLapComplete" whenever the player will run through a collider with a certain tag.
LapManager.cs is listening to "OnLapComplete" and will update the UI accordingly. It also checks if the player finishes all the laps and "wins the game" by calling "AllLapsComplete" - but right now it's just a log message.
I modified TimerController.cs so it calls "OnGameLoose" if the timer gets to 0 before you finish the laps and "looses the game" - still just a log message
Both for win and lost state the timer stops